+++ client/components/resource-details/ResourceInfo.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Resource, resourceLookup } from '../../utils'
 import { Gallery } from './Gallery'
 import { Reviews } from './Reviews'
-import { capitalizeFirstLetter } from '../common/Functions'
+import { capitalizeFirstLetter } from '../common/Utils'
 import { PageTitle, fontWeight } from '../../styles'
 import { colours, width, smallerThan } from '../../styles'
